+++ __brick__/{{project_name.snakeCase()}}/test/src/presentation/widgets/user_card_widget_test.dart
@@ -8,6 +8,8 @@ import 'package:mocktail/mocktail.dart';
 import '../mock_helpers.dart';
 import '../../entities/entity_helpers.dart';
 
+/// TODO: image network error test
+///
 void main() {
   late MockOnTap mockOnTap;