+++ __brick__/{{project_name.snakeCase()}}/test/src/repositories/user_repository_test.dart
@@ -1,6 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:{{project_name.snakeCase()}}/core/utils/failure.dart';
 import 'package:{{project_name.snakeCase()}}/src/data_sources/user_data_source.dart';
+import 'package:{{project_name.snakeCase()}}/src/data_sources/user_local_data_source.dart';
 import 'package:{{project_name.snakeCase()}}/src/repositories/user_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
@@ -9,14 +10,26 @@ import 'package:mockito/mockito.dart';
 import '../entities/entity_helpers.dart';
 import 'user_repository_test.mocks.dart';
 
-@GenerateMocks([UserDataSource])
+@GenerateMocks([
+  UserDataSource,
+  UserLocalDataSource,
+  // UserSqfliteDataSource,
+])
 void main() {
   late UserRepositoryImpl repository;
   late MockUserDataSource mockUserDataSource;
+  late MockUserLocalDataSource mockUserLocalDataSource;
+  // late MockUserSqfliteDataSource mockUserSqfliteDataSource;
 
   setUp(() {
     mockUserDataSource = MockUserDataSource();
-    repository = UserRepositoryImpl(dataSource: mockUserDataSource);
+    mockUserLocalDataSource = MockUserLocalDataSource();
+    // mockUserSqfliteDataSource = MockUserSqfliteDataSource();
+    repository = UserRepositoryImpl(
+      dataSource: mockUserDataSource,
+      localDataSource: mockUserLocalDataSource,
+      // sqfliteDataSource: mockUserSqfliteDataSource,
+    );
   });
 
   group('getUsers', () {
@@ -46,21 +59,55 @@ void main() {
       verify(mockUserDataSource.getUsers(page: 1, limit: 10));
     });
 
-    test('should return UnexpectedFailure(), when throw failure', () async {
+    test('should return LocalStorageFailure(), when local data source error',
+        () async {
       when(mockUserDataSource.getUsers(
         page: anyNamed('page'),
         limit: anyNamed('limit'),
-      )).thenThrow(const UnexpectedFailure());
+      )).thenAnswer((_) async => users);
+      when(mockUserLocalDataSource.setUsers(
+        users: anyNamed('users'),
+      )).thenThrow(const LocalStorageFailure());
 
       final result = await repository.getUsers(page: 1, limit: 10);
 
-      expect((result as Left).value, isA<UnexpectedFailure>());
+      expect((result as Left).value, isA<LocalStorageFailure>());
 
-      verify(mockUserDataSource.getUsers(page: 1, limit: 10));
+      verifyInOrder([
+        mockUserDataSource.getUsers(page: 1, limit: 10),
+        mockUserLocalDataSource.setUsers(users: users),
+      ]);
     });
 
-    test('should return list of users', () async {
+    test(
+        'should return InternetConnectionFailure(), when internet connection error and empty local data source',
+        () async {
       when(mockUserDataSource.getUsers(
+        page: anyNamed('page'),
+        limit: anyNamed('limit'),
+      )).thenThrow(const InternetConnectionFailure());
+      when(mockUserLocalDataSource.getUsers(
+        page: anyNamed('page'),
+        limit: anyNamed('limit'),
+      )).thenAnswer((_) async => []);
+
+      final result = await repository.getUsers(page: 1, limit: 10);
+
+      expect((result as Left).value, isA<InternetConnectionFailure>());
+
+      verifyInOrder([
+        mockUserDataSource.getUsers(page: 1, limit: 10),
+      ]);
+    });
+
+    test(
+        'should return list of users, when internet connection error and local data source has data',
+        () async {
+      when(mockUserDataSource.getUsers(
+        page: anyNamed('page'),
+        limit: anyNamed('limit'),
+      )).thenThrow(const InternetConnectionFailure());
+      when(mockUserLocalDataSource.getUsers(
         page: anyNamed('page'),
         limit: anyNamed('limit'),
       )).thenAnswer((_) async => users);
@@ -69,7 +116,28 @@ void main() {
 
       expect((result as Right).value, users);
 
-      verify(mockUserDataSource.getUsers(page: 1, limit: 10));
+      verifyInOrder([
+        mockUserDataSource.getUsers(page: 1, limit: 10),
+      ]);
+    });
+
+    test('should return list of users', () async {
+      when(mockUserDataSource.getUsers(
+        page: anyNamed('page'),
+        limit: anyNamed('limit'),
+      )).thenAnswer((_) async => users);
+      when(mockUserLocalDataSource.setUsers(
+        users: anyNamed('users'),
+      )).thenAnswer((_) async => unit);
+
+      final result = await repository.getUsers(page: 1, limit: 10);
+
+      expect((result as Right).value, users);
+
+      verifyInOrder([
+        mockUserDataSource.getUsers(page: 1, limit: 10),
+        mockUserLocalDataSource.setUsers(users: users),
+      ]);
     });
   });
 
@@ -94,6 +162,39 @@ void main() {
       final result = await repository.getUser(id: 'anyId');
 
       expect((result as Left).value, isA<UnexpectedFailure>());
+
+      verify(mockUserDataSource.getUser(id: 'anyId'));
+    });
+
+    test('should return LocalStorageFailure(), when local data source error',
+        () async {
+      when(mockUserDataSource.getUser(
+        id: anyNamed('id'),
+      )).thenAnswer((_) async => user);
+      when(mockUserLocalDataSource.setUser(
+        user: anyNamed('user'),
+      )).thenThrow(const LocalStorageFailure());
+
+      final result = await repository.getUser(id: 'anyId');
+
+      expect((result as Left).value, isA<LocalStorageFailure>());
+
+      verifyInOrder([
+        mockUserDataSource.getUser(id: 'anyId'),
+        mockUserLocalDataSource.setUser(user: user),
+      ]);
+    });
+
+    test(
+        'should return InternetConnectionFailure(), when internet connection error and empty local data source',
+        () async {
+      when(mockUserDataSource.getUser(
+        id: anyNamed('id'),
+      )).thenThrow(const InternetConnectionFailure());
+
+      final result = await repository.getUser(id: 'anyId');
+
+      expect((result as Left).value, isA<InternetConnectionFailure>());
 
       verify(mockUserDataSource.getUser(id: 'anyId'));
     });